% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildMultiBlock.R
\name{BuildMultiBlock}
\alias{BuildMultiBlock}
\title{BuildMultiBlock}
\usage{
BuildMultiBlock(
  newBlock = newBlock,
  newVars = NULL,
  newSamples = NULL,
  growingMB = NULL,
  batches = NULL,
  equalSampleNumber = T,
  ...
)
}
\arguments{
\item{newBlock}{The data to be incorporated into the multi-block structure.}

\item{newVars}{The variable names (optional)}

\item{newSamples}{The sample names. The sample replicates should have the same name (optional).}

\item{growingMB}{The name of the multi-block structure it this already exists (optional).}

\item{batches}{A vector indicating the batch number for every sample (optional).}

\item{equalSampleNumber}{Logical to indicate wether the number of samples is the same (T) or not (F) in all the blocks.}

\item{...}{Additional objects descriptive of newBlock can be introduced, as long as they have the same length as the sample size.}
}
\value{
The multi-block
}
\description{
Creates a multi-block structure compatible with ComDim_PCA_2020.R
}
\examples{
 b1 = matrix(rnorm(500),10,50) # 10 rows and 50 columns
 b2 = matrix(rnorm(800),10,80) # 10 rows and 80 columns
 b3 = matrix(rnorm(700),10,70) # 10 rows and 70 columns
 # Build multi-block by adding one data block at a time:
 mb <- BuildMultiBlock(b1, newSamples = paste0('sample_',1:10))
 mb <- BuildMultiBlock(b2, growingMB = mb)
 mb <- BuildMultiBlock(b3, growingMB = mb)
}
