% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitRB.R
\name{splitRB}
\alias{splitRB}
\title{splitRB - Building the RB list compatible with comdim_PCA().}
\usage{
splitRB(MB = MB, checkSampleCorrespondence = FALSE)
}
\arguments{
\item{MB}{The multi-block structure built with BuildMultiblock().}

\item{checkSampleCorrespondence}{If FALSE, the same number of samples and the sample order are assumed for all the batches. If TRUE, only the samples found in all replicate blocks will be included in the final structure.}
}
\value{
The list containing the multi-block composed of the replicate-blocks.
}
\description{
To split a multi-block into a list containing smaller blocks, comprising each data from one batch.
}
\examples{
b1 = matrix(rnorm(500),10,50)
batch_b1 = rep(1,10)
b2 = matrix(rnorm(800),30,80)
batch_b2 = c(rep(1,10),rep(2,10),rep(3,10))
# Generate the multi-block (mb)
mb <- BuildMultiBlock(b1, batches = batch_b1)   
mb <- BuildMultiBlock(b2, growingMB = mb, batches = batch_b2, equalSampleNumber = FALSE)
rb <- splitRB(mb)
}
